책 추가화면에 별점 표시 추가

+++ app/src/main/java/com/example/bookchigibakchigi/network/model/AladinBookItem.kt
@@ -22,7 +22,7 @@ data class AladinBookItem(
     val salesPoint: Int,
     val adult: Boolean,
     val fixedPrice: Boolean,
-    val customerReviewRank: Int,
+    val customerReviewRank: Float,
     val bestDuration: String?,
     val bestRank: Int?,
     val seriesInfo: SeriesInfo?,
